Initial commit for Autodetecting_Script
Working on making it so the production test can detect what type of AIO is hooked up and then be able to run the appropriate script, lowering the complexity of usage.

The original functionality would still exist; right now it's triggered by running a script with the tag name MasterScript, which is a placeholder name.

The production test can now completely identify which AIO is plugged in, including revision number and types of modules in each bay.

Next goal is to get it to be able to run pre-determined scripts based off these detections, such as AIO-2 Full Test.xml for an AIO 2.